feat: Add support to config for email via SMTP (`email-smtp` feature)
Also:

- Move contents of module `app_config` to parent `config` module
- Update `feature_powerset` feature groups
- Add `lettre` dependency for email/smtp support
- Add `cargo-insta` to `just init` command
- Add `--no-fail-fast` to `just test` sub-commands and `pre-push` hook
