+++ examples/full/build.rs
@@ -14,7 +14,7 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
         let out_dir = PathBuf::from(env::var("OUT_DIR").unwrap());
         tonic_build::configure()
             .file_descriptor_set_path(out_dir.join("helloworld_descriptor.bin"))
-            .compile(&["proto/helloworld.proto"], &["proto"])
+            .compile_protos(&["proto/helloworld.proto"], &["proto"])
             .unwrap();
     }
