+++ examples/leptos-ssr/src/server/mod.rs
@@ -11,8 +11,8 @@ use migration::Migrator;
 use roadster::app::context::AppContext;
 use roadster::app::metadata::AppMetadata;
 use roadster::app::App as RoadsterApp;
-use roadster::config::app_config::AppConfig;
 use roadster::config::environment::Environment;
+use roadster::config::AppConfig;
 use roadster::error::RoadsterResult;
 use roadster::service::http::service::HttpService;
 use roadster::service::registry::ServiceRegistry;
