+++ src/api/http/health.rs
@@ -149,7 +149,7 @@ fn health_get_docs(op: TransformOperation) -> TransformOperation {
 #[cfg(test)]
 mod tests {
     use crate::app::context::AppContext;
-    use crate::config::app_config::AppConfig;
+    use crate::config::AppConfig;
     use rstest::rstest;
 
     // Todo: Is there a better way to structure this test (and the ones in `docs` and `ping`)
