+++ src/app/roadster_app.rs
@@ -4,7 +4,7 @@ use crate::app;
 use crate::app::context::AppContext;
 use crate::app::metadata::AppMetadata;
 use crate::app::App;
-use crate::config::app_config::AppConfig;
+use crate::config::AppConfig;
 use crate::error::RoadsterResult;
 use crate::health_check::registry::HealthCheckRegistry;
 use crate::health_check::HealthCheck;
