+++ src/config/service/http/default_routes.rs
@@ -82,8 +82,8 @@ impl DefaultRouteConfig {
 
 #[cfg(test)]
 mod tests {
-    use crate::config::app_config::AppConfig;
     use crate::config::service::http::*;
+    use crate::config::AppConfig;
     use rstest::rstest;
 
     #[rstest]
