+++ src/config/service/http/initializer.rs
@@ -1,5 +1,5 @@
 use crate::app::context::AppContext;
-use crate::config::app_config::CustomConfig;
+use crate::config::CustomConfig;
 use crate::service::http::initializer::normalize_path::NormalizePathConfig;
 use crate::util::serde::default_true;
 use axum::extract::FromRef;
