+++ src/health_check/database.rs
@@ -37,7 +37,7 @@ fn enabled(context: &AppContext) -> bool {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::config::app_config::AppConfig;
+    use crate::config::AppConfig;
     use rstest::rstest;
 
     #[rstest]
