+++ src/service/http/middleware/compression.rs
@@ -105,7 +105,7 @@ where
 mod tests {
     use super::*;
     use crate::app::context::AppContext;
-    use crate::config::app_config::AppConfig;
+    use crate::config::AppConfig;
     use rstest::rstest;
 
     #[rstest]
