+++ src/service/http/middleware/default.rs
@@ -45,7 +45,7 @@ where
 #[cfg(test)]
 mod tests {
     use crate::app::context::AppContext;
-    use crate::config::app_config::AppConfig;
+    use crate::config::AppConfig;
     use crate::testing::snapshot::TestCase;
     use insta::assert_toml_snapshot;
     use itertools::Itertools;
