+++ src/service/http/middleware/request_id.rs
@@ -156,7 +156,7 @@ where
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::config::app_config::AppConfig;
+    use crate::config::AppConfig;
     use rstest::rstest;
 
     #[rstest]
