+++ src/service/http/middleware/sensitive_headers.rs
@@ -167,7 +167,7 @@ where
 mod tests {
     use super::*;
     use crate::app::context::AppContext;
-    use crate::config::app_config::AppConfig;
+    use crate::config::AppConfig;
     use rstest::rstest;
 
     #[rstest]
