+++ src/service/worker/sidekiq/builder.rs
@@ -295,7 +295,7 @@ where
 mod tests {
     use super::*;
     use crate::app::context::AppContext;
-    use crate::config::app_config::AppConfig;
+    use crate::config::AppConfig;
     use crate::service::worker::sidekiq::MockProcessor;
     use bb8::Pool;
     use futures::StreamExt;
